requirements.txt
beautifulsoup4==4.13.4
numpy==2.3.1
opencv_python==4.11.0.86
Pillow==11.3.0
Requests==2.32.4
selenium==4.33.0
undetected_chromedriver==3.5.5
